customer-portal-backend/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
Pillow==10.1.0
qrcode[pil]==7.4.2
python-dotenv==1.0.0
django-cors-headers==4.3.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
